+++ modules/canvaskit/gpu.js
@@ -156,15 +156,11 @@
 
       function pushTexture(tex) {
         // GL is an emscripten object that holds onto WebGL state. One item in that state is
-        // an array of textures, of which the index is the handle/id.
-        var texHandle = GL.textures.length;
-        if (!texHandle) {
-          // If our texture handle is 0, Skia interprets that as an invalid texture id.
-          // As a special case, we push a null texture there so the first texture has id 1.
-          GL.textures.push(null);
-          texHandle = 1;
-        }
-        GL.textures.push(tex);
+        // an array of textures, of which the index is the handle/id. We must call getNewId so
+        // the GL's tracking of textures is up to date and we do not accidentally use the same
+        // texture in two different places if Skia creates a texture. (e.g. skbug.com/12797)
+        var texHandle = GL.getNewId(GL.textures);
+        GL.textures[texHandle] = tex;
         return texHandle
       }